feat: Make _SliderMenu public
