+++ lib/selectable_text/screen/selectable_text_screen.dart
@@ -8,6 +8,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intersperse/intersperse.dart';
 
 import 'package:dismantling/components/select_button.dart';
+import 'package:dismantling/components/slider_menu.dart';
 import 'package:dismantling/components/toggle_button.dart';
 import 'package:dismantling/selectable_text/view_model/selectable_text_view_model.dart';
 
@@ -72,7 +73,7 @@ class SelectableTextScreen extends HookConsumerWidget {
                           onToggled: (value) => notifier.showCursorToggled(),
                         ),
                         if (state.showCursor) ...[
-                          _SliderMenu(
+                          SliderMenu(
                             label: 'cursorWidth',
                             value: state.cursorWidth,
                             min: 1,
@@ -80,7 +81,7 @@ class SelectableTextScreen extends HookConsumerWidget {
                             divisions: 19,
                             onChanged: notifier.cursorWidthUpdated,
                           ),
-                          _SliderMenu(
+                          SliderMenu(
                             label: 'cursorHeight',
                             value: state.cursorHeight ?? 1,
                             min: 1,
@@ -89,7 +90,7 @@ class SelectableTextScreen extends HookConsumerWidget {
                             onChanged: notifier.cursorHeightUpdated,
                           ),
                           // TODO(ashdik): Add Radius.elliptical version
-                          _SliderMenu(
+                          SliderMenu(
                             label: 'cursorRadius',
                             value: state.cursorRadius?.x ?? 0,
                             min: 0,
@@ -192,62 +193,6 @@ class SelectableTextScreen extends HookConsumerWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class _SliderMenu extends StatelessWidget {
-  const _SliderMenu({
-    required this.label,
-    required this.value,
-    required this.min,
-    required this.max,
-    required this.divisions,
-    required this.onChanged,
-  });
-
-  final String label;
-  final double value;
-  final double min;
-  final double max;
-  final int divisions;
-
-  final ValueChanged<double> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Text(
-          label,
-          style: theme.textTheme.titleLarge!.copyWith(
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        const SizedBox(height: 20),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(min.toString(), style: theme.textTheme.bodySmall),
-            Text(
-              value.toString(),
-              style: theme.textTheme.titleLarge!.copyWith(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            Text(max.toString(), style: theme.textTheme.bodySmall),
-          ],
-        ),
-        Slider.adaptive(
-          value: value,
-          min: min,
-          max: max,
-          divisions: divisions,
-          onChanged: onChanged,
-        ),
-      ],
     );
   }
 }
